move component layout under components
